work in progress establish OsmElementController between element and geometry DAO and rest of the app

+++ app/src/main/java/de/westnordost/streetcomplete/data/osm/osmquest/OsmQuestDao.kt
@@ -85,6 +85,7 @@ internal class OsmQuestDao @Inject constructor(
     }
 
     fun deleteAllIds(ids: Collection<Long>): Int {
+        if (ids.isEmpty()) return 0
         return db.delete(NAME, "$QUEST_ID IN (${ids.joinToString(",")})", null)
     }
